# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EarScope API",
        "description": "Dokumentasi API untuk EarScope",
        "version": "1.0"
    },
    "paths": {
...
        "/v1/register/pasien": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register a new pasien",
                "operationId": "1d0c0ab5739307e0b25ebbd038c11254",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "tanggal_lahir",
                                    "gender",
                                    "no_telp"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "tanggal_lahir": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2012-12-31"
                                    },
                                    "gender": {
                                        "description": "Jenis kelamin, hanya diisi 'laki-laki' atau 'perempuan'",
                                        "type": "string",
                                        "example": "laki-laki"
                                    },
                                    "no_telp": {
                                        "description": "Nomor telepon, hanya angka, 10 sampai 15 digit",
                                        "type": "string",
                                        "example": "081234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Nama sudah terdaftar, silakan gunakan nama lain."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/register/dokter": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register a new dokter",
                "operationId": "67328887bff28886787b4af930d4bacd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "tanggal_lahir",
                                    "gender",
                                    "no_telp"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "tanggal_lahir": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2012-12-31"
                                    },
                                    "gender": {
                                        "description": "Jenis kelamin, hanya diisi 'laki-laki' atau 'perempuan'",
                                        "type": "string",
                                        "example": "laki-laki"
                                    },
                                    "no_telp": {
                                        "description": "Nomor telepon, hanya angka, 10 sampai 15 digit",
                                        "type": "string",
                                        "example": "081234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Nama sudah terdaftar, silakan gunakan nama lain."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get authenticated user",
                "operationId": "b32adad160e96976779040f9684ccb22",
                "responses": {
                    "200": {
                        "description": "Authenticated user data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/v1/users/{id}/force-delete": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Permanently delete a user",
                "operationId": "83015cd4500e57fea8350f9a726e60d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User permanently deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/users/{id}/restore": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Restore a soft deleted user",
                "operationId": "74ef9272bd28fb028e3225ad06f44cbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User restored successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Masukkan token di kolom ini dengan format Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}